reflexion-agent/pyproject.toml
[project]
name = "reflexion-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "black>=25.11.0",
    "isort>=6.1.0",
    "langchain>=0.3.27",
    "langchain-aws>=0.2.35",
    "langchain-community>=0.3.31",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
